requirements.txt
alembic==1.10.4
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.1
colorama==0.4.6
colorlib==0.1.0
contourpy==1.0.6
cryptography==39.0.1
cssselect2==0.7.0
cycler==0.11.0
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.5
django-allauth==0.52.0
django-classy-tags==4.0.0
django-cms==3.11.1
django-cors-headers==3.14.0
django-crispy-forms==1.14.0
django-formtools==2.4
django-heroku==0.3.1
django-jazzmin==2.6.0
django-model-utils==4.3.1
django-sekizai==4.0.0
django-treebeard==4.4
django-widget-tweaks==1.4.12
djangocms-admin-style==3.2.3
djangocms-ads==1.1.0
djangocms-attributes-field==2.1.0
djangorestframework==3.14.0
filelock==3.9.0
Flask==2.1.0
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
fonttools==4.38.0
greenlet==2.0.2
gunicorn==20.1.0
html5lib==1.1
idna==3.4
itsdangerous==2.0.1
Jinja2==3.0.1
keras==2.11.0
kiwisolver==1.4.4
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.0.1
matplotlib==3.6.3
mysql-connector-python==8.0.26
mysqlclient==2.1.1
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.7.0.68
oscrypto==1.3.0
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
pipenv==2023.2.4
platformdirs==2.6.2
protobuf==4.22.3
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyAudio==0.2.13
pycparser==2.21
pydub==0.25.1
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
PyJWT==2.6.0
pyparsing==3.0.9
pypdf==3.6.0
pyperclip3==0.4.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pywin32==305
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
scipy==1.10.0
six==1.16.0
sklearn==0.0.post1
SpeechRecognition==3.9.0
SQLAlchemy==1.4.23
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
uritools==4.0.1
urllib3==1.26.14
virtualenv==20.17.1
virtualenv-clone==0.5.7
webencodings==0.5.1
Werkzeug==2.0.1
whitenoise==6.3.0
WTForms==2.3.3
xhtml2pdf==0.2.9
